tqqq_bot_aws_ec2/requirements.txt
# Core Alpaca Trading API
alpaca-py>=0.13.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Timezone handling
pytz>=2023.3

# Environment variables
python-dotenv>=1.0.0

# Desktop notifications (optional but recommended)
plyer>=2.1.0

# Logging enhancements
colorlog>=6.7.0

# AWS SDK (for EC2 deployment)
boto3>=1.26.0

# Scheduling (for automated runs)
schedule>=1.2.0

# System monitoring
psutil>=5.9.0

# HTTP server for health checks
flask>=2.3.0

# Data visualization (optional, for analysis)
matplotlib>=3.7.0
plotly>=5.14.0

# Testing
pytest>=7.3.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
